+++ src/PyReprism/languages/coffeescript.py
@@ -17,7 +17,7 @@ class CoffeeScript(BaseLanguage):
 
     @classmethod
     def comment_regex(cls):
-        return re.compile(r'(?P<comment>#.*?$|###.*?###)|(?P<noncomment>[^#\n].*?$)', re.DOTALL | re.MULTILINE)
+        return re.compile(r'(?P<comment>###[\s\S]*?###|#.*?$)|(?P<noncomment>.[^#]*)', re.DOTALL | re.MULTILINE)
 
     @classmethod
     def number_regex(cls):